restdjagnocurd/curd_rest_django/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
breadability==0.1.20
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.3.2
cassandra-pylib==0.0.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.4
command-not-found==0.3
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
djangorestframework==3.15.1
docopt==0.6.2
duplicity==0.8.21
emoji==1.7.0
etelemetry==0.3.1
exceptiongroup==1.2.0
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.50.0
future==0.18.2
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.59.2
gyp==0.1
h11==0.14.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.6.4
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
leadtools==22.0.0.8
leadtools-common-runtime==22.0.0.8
libclang==16.0.6
lockfile==0.12.2
looseversion==1.3.0
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.3
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
openai==1.11.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
pathlib==1.0.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==4.25.0
prov==2.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycountry==23.12.11
pycparser==2.21
pycups==2.0.1
pydantic==2.6.0
pydantic_core==2.16.1
pydot==2.0.0
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyNaCl==1.5.0
pynmea2==1.19.0
pyparsing==3.1.1
pyRFC3339==1.1
PySocks==1.7.1
pyspark==3.5.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pythonnet==3.0.3
pytz==2022.1
pyxdg==0.27
pyxnat==1.6.2
PyYAML==5.4.1
rdflib==7.0.0
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.12.0
sdkmanager==0.5.1
SecretStorage==3.3.1
selenium==4.17.2
serial==0.0.97
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
ssh-import-id==5.11
sumy==0.11.0
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
text2emotion==0.0.5
tk==0.1.0
tqdm==4.66.1
traits==6.3.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vaderSentiment==3.3.2
vboxapi==1.0
virtualenv==20.25.1
wadllib==1.3.6
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
